configs: bsp.ref: ignore unused ModemManager libraries
Our internal reason-tool checks for unused libraries in the rootfs.
ModemManager is used on the target device und shall support multiple
protocols. Before we optimize which libraries are actually needed, we
leave it at the default and ignore that some libraries currently don't
have any tools that depend on them.

Once it is clear what is really needed, this list needs to be revisited.

Signed-off-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>