tests: debugging: enable the PSCI checker for configuration error detection
This SoC has two CPU cores and thus, the SMP_ON_UP should be disabled to
enforce the presence of the bootloader's PSCI feature.

This kernel must be started in 'nonsecure' mode, because the bootloader
must be used by the kernel to hotplug CPU cores. This checker tests the
corresponding PSCI features the bootloader provides and is just a
development tool. It should be kept enabled until the issues with the
second CPU core are solved.

After fixing the issue, both the 'SMP_ON_UP' and 'ARM_PSCI_CHECKER' should be
disabled again.