gateway: make /config a real, nonempty partition
Until now, /config was just an empty partition. To have in in the final
image with the real size, we had to fake it with an empty directory.
Resizing/Growing on the hardware doesn't work with extendend partitions.

Now /config is a real image where files can be installed into.

Signed-off-by: Steffen Trumtrar <s.trumtrar@pengutronix.de>